agents/autogen/mcp_agent/mcp_automl_template/pyproject.toml
[project]
name = "mcp-automl-template"
version = "0.1.0"
description = "MCP server with config-driven tools (JSON schema + YAML). Optional HTTP deployment (e.g. AutoML/KServe) for tools. SSE transport for Cursor, Ollama, and agents."
authors = [{ name = "Your Name", email = "you@example.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "langchain-openai>=1.1.10,<2.0.0",
    "langchain>=1.2.10,<2.0.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0,<1.0.0",
    "langgraph>=1.0.0",
    "python-dotenv>=1.2.0,<2.0.0",
    "mcp>=1.0.0",
    "langchain_mcp_adapters>=0.1.0,<1.0.0",
    "langchain-llama-stack>=0.1.0,<1.0.0",
    "pandas>=2.0.0,<3.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "uvicorn[standard]>=0.41.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2,<10.0.0",
]

[tool.setuptools]
py-modules = ["mcp_server", "register_tools", "utils"]

[build-system]
requires = ["setuptools>=80.9.0"]
build-backend = "setuptools.build_meta"
